Making sure that AL_ORGANIZATION_TITLE is defined
We have been running into the following error on prod for a while:

```
DAErrorMissingVariable: 

    Interview has an error.  There was a reference to a variable 'AL_ORGANIZATION_TITLE' that could not be looked up in the question file (for language 'en') or in any of the files incorporated by reference into the question file.

The referer URL was http://suffolklitlab.org
```
But it's been impossible to find where it was happening from the email. I tried at one point opening and going through every single interview on court forms online, since the referer URL was suffolklitlab.org, but none of them failed on the landing page, which is what I would have expected from that referer.

Finally realized that the DA logs might have more information, and I was able to find that it's coming from this file, which some interview (still don't know which) is including. Including the `al_settings.yml` file should fix it, and have `AL_ORGANIZATION_TITLE` be defined now.