Make audo buttons have same names as aria labels
"When speech input users interact with a web page, they usually speak
a command followed by the reference to some visible label (like the
text in a button or a link, or the text labelling some input). If the
visible label does not match the accessible name of the control, speech
users may be unable to directly activate that control."

The options are to remove "Listen" (bad for non-assistive tech users IMO),
make all of the aria-labels "Listen" (bad for assistive tech users),
or just make the visible text match the aria-labels.

https://www.w3.org/WAI/WCAG21/Techniques/failures/F96