Add back auto-ocr functionality and bates numbers
To the ALExhibitDocument. Main purpose is so we can update the
MotionToStayEviction interview to use AL Core instead of MAVC, so it can use
e-filing.

I moved some of the functions to the ALExhibit itself to make things easier to
read, and added the logic to add bates numbers, keeping the right numbers
with table of contents and cover pages.