Add focus shadow for back button
Before this change, you aren't able to see that you are keyboard tabbing
over the back button, as the value bootstrap uses to color the shadow
doesn't have a value for the `daquestionbackbutton` class, only
the `btn-primary` class.

TBH, I'm not a huge fan of the default focus indicators in bootstrap,
they don't have enough contrast. But that will likely have to wait
for a different PR, as it's a much larger change.