Speed up next screen after uploading Exhibit page
By calling `reset_gathered()` and `gather()` on the ALExhibitList, we force
DA to call `str` on each document we upload. We can avoid this by setting the
`complete_attribute` to `ok`, which is always defined for a DAFile.

Thumbnails are still generated when the PDF is uploaded, but it happens in the
background, and won't cause the 504 sometimes seen.

Partially fix #752.