Handle unexpected interview_metadata
Ran into an error when `interview_metadat["main_interview_key"}` was set to
the full metadata dictionary. Definitely not expected use, but it shouldn't
error out if that's the case.

This PR adds a check to make sure the interview key is a string like we expect
it to be. If it's not, we fallback to the other methods (just checking the
only entry in `interview_metadata` like we do in the next check works in this
case).