Include AssemblyLine through interview list
At MLH, we make a `interview_list` wrapper recently, that just included our
interview framework YAML and `interview_list.yml`.

However, in doing so, `al_audio.js` is added twice to the page twice, and runs
twice, so there are two widgets that show up.

IMO, `interview_list.yml` should just use AssemblyLine fully, instead of the
awkward half-includes that it currently does. The only changes that happen when
you do so are the red `pre` debug text and the footer, both of which I manually
set to empty. People are welcome to re-enable those on `interview_list` if they
wish.

I'm working an a PR to stop the JS stuff from running twice, even if it's on
the page twice, still in progress there.