International addresses (#794)
Minor changes that address #281. More details in that issue, but I don't think it's possible to get the full state dropdown for each country without messing up Google autocomplete, so when the country is an input, we turn the state list into a text box (I tried swapping out a dropdown for just the US and text input for other countries, but Google autocomplete also has a problem with that).

Only other changes were to change the "State" label to "State / Province" when the country is an input, and a few comments / docs, for accuracy.