Fixed failing mock functions in language test
Because `unittest` no longer works with PEP-420 implicit namespaces
(which we are required to use because of docassemble and other python
tooling deprecating the old way of using namespaces), we run unit tests
with `python -m unittest discover docassemble`.

Starting at the `docassemble` folder changes what `unittest` thinks
the namespace is, so the `patch` call needs to patch the `AssemblyLine.language.url_action`
function instead. We keep the original patched function (`docassemble.AssemblyLine...`)
to make it clear what we are trying to override and for the tests to work
when run without the discover feature.

We also change to using `patch` as a decorator, which is it's documented
use, and more common to find examples using, even though it has to be
repeated for each test function.